+++ test/escape.spec.ts
@@ -4,23 +4,20 @@ import { bibtex, bibtexTidy, test } from "./utils";
 const input = bibtex`
 @misc{q,
   author = {Chars \", _, $, @, &, é, ɛ, ū, {"}},
-  title = {Math expression $\\alpha$ stay unescaped},
-  url = {something#boo}
+  title = {Math expression $\\alpha$ stay unescaped}
 }`;
 
 const outputEscaped = bibtex`
 @misc{q,
   author        = {Chars \", \_, \$, \@, \&, \'{e}, \varepsilon{}, \={u}, {"}},
-  title         = {Math expression $\\alpha$ stay unescaped},
-  url           = {something\#boo}
+  title         = {Math expression $\\alpha$ stay unescaped}
 }
 `;
 
 const outputUnescaped = bibtex`
 @misc{q,
   author        = {Chars \", _, $, @, &, é, ɛ, ū, {"}},
-  title         = {Math expression $\\alpha$ stay unescaped},
-  url           = {something#boo}
+  title         = {Math expression $\\alpha$ stay unescaped}
 }
 `;