+++ CHANGELOG.md
@@ -184,3 +184,4 @@ Providing multiple input files will only work in `--modify`/`-m` mode.
 - Do not flatten braces which contain commands (#407, #423)
 - Do not encode special characters in verbatim fields like url (#415)
 - Upgrade dependencies
+- Pre-commit hook changed to script so that dev dependencies are not required (#427)