+++ CHANGELOG.md
@@ -177,3 +177,11 @@ Providing multiple input files will only work in `--modify`/`-m` mode.
 - Upgrade deps
 - Accessibility tweak
 - Upgrade dependencies
+
+### Unrelease
+
+- Support filenames with spaces (#416, #428)
+- Do not flatten braces which contain commands (#407, #423)
+- Do not encode special characters in verbatim fields like url (#415)
+- Upgrade dependencies
+- Pre-commit hook changed to script so that dev dependencies are not required (#427)